Fix FLAC decoded size calculation and 64-bit seek truncation
Use sizeof(int16_t) for FLAC filesize/frame calculations since
drflac_read_pcm_frames_s16 always outputs 16-bit samples, regardless
of source bit depth (e.g. 24-bit FLAC would overallocate).

Widen posixemu_seek and io_lseek from int to off_t so that seek offsets
on large disk images (>2GB HDF/CHD) are not truncated to 32 bits.

+++ src/include/sysdeps.h
@@ -289,7 +289,7 @@ extern int posixemu_close (int);
 extern int posixemu_read (int, void *, int);
 extern int posixemu_write (int, const void *, int);
 #undef lseek
-extern int posixemu_seek (int, int, int);
+extern off_t posixemu_seek (int, off_t, int);
 extern int posixemu_stat (const TCHAR *, STAT *);
 extern int posixemu_mkdir (const TCHAR *, int);
 extern int posixemu_rmdir (const TCHAR *);
